move apis to the registry part
