+++ pkg/api/playlist.go
@@ -12,8 +12,8 @@ import (
 	"github.com/grafana/grafana/pkg/api/dtos"
 	"github.com/grafana/grafana/pkg/api/response"
 	"github.com/grafana/grafana/pkg/api/routing"
-	internalplaylist "github.com/grafana/grafana/pkg/apis/playlist"
 	"github.com/grafana/grafana/pkg/middleware"
+	internalplaylist "github.com/grafana/grafana/pkg/registry/apis/playlist"
 	contextmodel "github.com/grafana/grafana/pkg/services/contexthandler/model"
 	"github.com/grafana/grafana/pkg/services/featuremgmt"
 	"github.com/grafana/grafana/pkg/services/grafana-apiserver/endpoints/request"
