--- pkg/apis/playlist/register.go
@@ -1,118 +1,0 @@
-package playlist
-
-import (
-	"fmt"
-	"time"
-
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/runtime/schema"
-	"k8s.io/apimachinery/pkg/runtime/serializer"
-	"k8s.io/apiserver/pkg/registry/generic"
-	"k8s.io/apiserver/pkg/registry/rest"
-	genericapiserver "k8s.io/apiserver/pkg/server"
-	common "k8s.io/kube-openapi/pkg/common"
-
-	grafanaapiserver "github.com/grafana/grafana/pkg/services/grafana-apiserver"
-	"github.com/grafana/grafana/pkg/services/grafana-apiserver/endpoints/request"
-	grafanarest "github.com/grafana/grafana/pkg/services/grafana-apiserver/rest"
-	"github.com/grafana/grafana/pkg/services/grafana-apiserver/utils"
-	"github.com/grafana/grafana/pkg/services/playlist"
-	"github.com/grafana/grafana/pkg/setting"
-)
-
-// GroupName is the group name for this API.
-const GroupName = "playlist.grafana.app"
-const VersionID = runtime.APIVersionInternal
-
-var _ grafanaapiserver.APIGroupBuilder = (*PlaylistAPIBuilder)(nil)
-
-// This is used just so wire has something unique to return
-type PlaylistAPIBuilder struct {
-	service    playlist.Service
-	namespacer request.NamespaceMapper
-	gv         schema.GroupVersion
-}
-
-func RegisterAPIService(p playlist.Service,
-	apiregistration grafanaapiserver.APIRegistrar,
-	cfg *setting.Cfg,
-) *PlaylistAPIBuilder {
-	builder := &PlaylistAPIBuilder{
-		service:    p,
-		namespacer: request.GetNamespaceMapper(cfg),
-		gv:         schema.GroupVersion{Group: GroupName, Version: VersionID},
-	}
-	apiregistration.RegisterAPI(builder)
-	return builder
-}
-
-func (b *PlaylistAPIBuilder) GetGroupVersion() schema.GroupVersion {
-	return b.gv
-}
-
-func (b *PlaylistAPIBuilder) InstallSchema(scheme *runtime.Scheme) error {
-	scheme.AddKnownTypes(b.gv,
-		&Playlist{},
-		&PlaylistList{},
-	)
-	return nil
-}
-
-func (b *PlaylistAPIBuilder) GetAPIGroupInfo(
-	scheme *runtime.Scheme,
-	codecs serializer.CodecFactory, // pointer?
-	optsGetter generic.RESTOptionsGetter,
-) (*genericapiserver.APIGroupInfo, error) {
-	apiGroupInfo := genericapiserver.NewDefaultAPIGroupInfo(GroupName, scheme, metav1.ParameterCodec, codecs)
-	storage := map[string]rest.Storage{}
-
-	legacyStore := &legacyStorage{
-		service:                   b.service,
-		namespacer:                b.namespacer,
-		DefaultQualifiedResource:  b.gv.WithResource("playlists").GroupResource(),
-		SingularQualifiedResource: b.gv.WithResource("playlist").GroupResource(),
-	}
-	legacyStore.tableConverter = utils.NewTableConverter(
-		legacyStore.DefaultQualifiedResource,
-		[]metav1.TableColumnDefinition{
-			{Name: "Name", Type: "string", Format: "name"},
-			{Name: "Title", Type: "string", Format: "string", Description: "The playlist name"},
-			{Name: "Interval", Type: "string", Format: "string", Description: "How often the playlist will update"},
-			{Name: "Created At", Type: "date"},
-		},
-		func(obj runtime.Object) ([]interface{}, error) {
-			m, ok := obj.(*Playlist)
-			if !ok {
-				return nil, fmt.Errorf("expected playlist")
-			}
-			return []interface{}{
-				m.Name,
-				m.Spec.Title,
-				m.Spec.Interval,
-				m.CreationTimestamp.UTC().Format(time.RFC3339),
-			}, nil
-		},
-	)
-	storage["playlists"] = legacyStore
-
-	// enable dual writes if a RESTOptionsGetter is provided
-	if optsGetter != nil {
-		store, err := newStorage(scheme, optsGetter, legacyStore)
-		if err != nil {
-			return nil, err
-		}
-		storage["playlists"] = grafanarest.NewDualWriter(legacyStore, store)
-	}
-
-	apiGroupInfo.VersionedResourcesStorageMap["v0alpha1"] = storage
-	return &apiGroupInfo, nil
-}
-
-func (b *PlaylistAPIBuilder) GetOpenAPIDefinitions() common.GetOpenAPIDefinitions {
-	return nil // no custom OpenAPI definitions
-}
-
-func (b *PlaylistAPIBuilder) GetAPIRoutes() *grafanaapiserver.APIRoutes {
-	return nil // no custom API routes
-}
