--- pkg/apis/playlist/v0alpha1/register.go
@@ -1,74 +1,0 @@
-package v0alpha1
-
-import (
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/runtime/schema"
-	"k8s.io/apimachinery/pkg/runtime/serializer"
-	"k8s.io/apiserver/pkg/registry/generic"
-	genericapiserver "k8s.io/apiserver/pkg/server"
-	common "k8s.io/kube-openapi/pkg/common"
-
-	grafanaapiserver "github.com/grafana/grafana/pkg/services/grafana-apiserver"
-	"github.com/grafana/grafana/pkg/services/grafana-apiserver/endpoints/request"
-	"github.com/grafana/grafana/pkg/services/playlist"
-	"github.com/grafana/grafana/pkg/setting"
-)
-
-// GroupName is the group name for this API.
-const GroupName = "playlist.grafana.app"
-const VersionID = "v0alpha1"
-
-var _ grafanaapiserver.APIGroupBuilder = (*PlaylistAPIBuilder)(nil)
-
-// This is used just so wire has something unique to return
-type PlaylistAPIBuilder struct {
-	service    playlist.Service
-	namespacer request.NamespaceMapper
-	gv         schema.GroupVersion
-}
-
-func RegisterAPIService(p playlist.Service,
-	apiregistration grafanaapiserver.APIRegistrar,
-	cfg *setting.Cfg,
-) *PlaylistAPIBuilder {
-	builder := &PlaylistAPIBuilder{
-		service:    p,
-		namespacer: request.GetNamespaceMapper(cfg),
-		gv:         schema.GroupVersion{Group: GroupName, Version: VersionID},
-	}
-	apiregistration.RegisterAPI(builder)
-	return builder
-}
-
-func (b *PlaylistAPIBuilder) GetGroupVersion() schema.GroupVersion {
-	return b.gv
-}
-
-func (b *PlaylistAPIBuilder) InstallSchema(scheme *runtime.Scheme) error {
-	scheme.AddKnownTypes(b.gv,
-		&Playlist{},
-		&PlaylistList{},
-	)
-	if err := RegisterConversions(scheme); err != nil {
-		return err
-	}
-	metav1.AddToGroupVersion(scheme, b.gv)
-	return scheme.SetVersionPriority(b.gv)
-}
-
-func (b *PlaylistAPIBuilder) GetAPIGroupInfo(
-	scheme *runtime.Scheme,
-	codecs serializer.CodecFactory, // pointer?
-	optsGetter generic.RESTOptionsGetter,
-) (*genericapiserver.APIGroupInfo, error) {
-	return nil, nil
-}
-
-func (b *PlaylistAPIBuilder) GetOpenAPIDefinitions() common.GetOpenAPIDefinitions {
-	return getOpenAPIDefinitions
-}
-
-func (b *PlaylistAPIBuilder) GetAPIRoutes() *grafanaapiserver.APIRoutes {
-	return nil // no custom API routes
-}
