--- pkg/apis/wireset.go
@@ -1,17 +1,0 @@
-package apis
-
-import (
-	"github.com/google/wire"
-
-	examplev0alpha1 "github.com/grafana/grafana/pkg/apis/example/v0alpha1"
-	"github.com/grafana/grafana/pkg/apis/playlist"
-	playlistsv0alpha1 "github.com/grafana/grafana/pkg/apis/playlist/v0alpha1"
-)
-
-// WireSet is the list of all services
-// NOTE: you must also register the service in: pkg/registry/apis/apis.go
-var WireSet = wire.NewSet(
-	playlist.RegisterAPIService,
-	playlistsv0alpha1.RegisterAPIService,
-	examplev0alpha1.RegisterAPIService,
-)
