+++ pkg/registry/apis/apis.go
@@ -3,10 +3,10 @@ package apiregistry
 import (
 	"context"
 
-	examplev0alpha1 "github.com/grafana/grafana/pkg/apis/example/v0alpha1"
-	"github.com/grafana/grafana/pkg/apis/playlist"
-	playlistsv0alpha1 "github.com/grafana/grafana/pkg/apis/playlist/v0alpha1"
 	"github.com/grafana/grafana/pkg/registry"
+	"github.com/grafana/grafana/pkg/registry/apis/example"
+	"github.com/grafana/grafana/pkg/registry/apis/playlist"
+	playlistV0 "github.com/grafana/grafana/pkg/registry/apis/playlist/v0alpha1"
 )
 
 var (
@@ -15,12 +15,12 @@ var (
 
 type Service struct{}
 
-// ProvideService is an entry point for each service that will force initialization
+// ProvideRegistryServiceSink is an entry point for each service that will force initialization
 // and give each builder the chance to register itself with the main server
-func ProvideService(
+func ProvideRegistryServiceSink(
+	_ *playlistV0.PlaylistAPIBuilder,
 	_ *playlist.PlaylistAPIBuilder,
-	_ *playlistsv0alpha1.PlaylistAPIBuilder,
-	_ *examplev0alpha1.TestingAPIBuilder,
+	_ *example.TestingAPIBuilder,
 ) *Service {
 	return &Service{}
 }
