+++ pkg/services/grafana-apiserver/service.go
@@ -307,7 +307,7 @@ func (s *service) start(ctx context.Context) error {
 		if err != nil {
 			return err
 		}
-		if g == nil {
+		if g == nil || len(g.PrioritizedVersions) < 1 {
 			continue
 		}
 		err = server.InstallAPIGroup(g)
